tests/wpt/web-platform-tests/webusb/usbConnectionEvent.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/fake-devices.js"></script>
<script src="resources/usb-helpers.js"></script>
<script>
'use strict';

usb_test(() => {
  return getFakeDevice().then(({ device }) => {
    let evt = new USBConnectionEvent('connect', { device: device });
    assert_equals(evt.type, 'connect');
    assert_equals(evt.device, device);
  });
}, 'Can construct a USBConnectionEvent with a device');

test(t => {
  assert_throws(TypeError(), () =>
      new USBConnectionEvent('connect', { device: null }));
  assert_throws(TypeError(), () => new USBConnectionEvent('connect', {}));
}, 'Cannot construct a USBConnectionEvent without a device');
</script>
